+++ packages/github/github_server.js
@@ -8,7 +8,7 @@ OAuth.registerService('github', 2, null, function(query) {
   return {
     serviceData: {
       id: identity.id,
-      accessToken: {seal: accessToken},
+      accessToken: OAuth.sealSecret(accessToken),
       email: identity.email,
       username: identity.login
     },
@@ -37,7 +37,7 @@ var getAccessToken = function (query) {
         params: {
           code: query.code,
           client_id: config.clientId,
-          client_secret: OAuth._openSecret(config.secret),
+          client_secret: OAuth.openSecret(config.secret),
           redirect_uri: Meteor.absoluteUrl("_oauth/github?close"),
           state: query.state
         }