+++ packages/weibo/weibo_server.js
@@ -41,7 +41,7 @@ var getTokenResponse = function (query) {
       "https://api.weibo.com/oauth2/access_token", {params: {
         code: query.code,
         client_id: config.clientId,
-        client_secret: OAuth._openSecret(config.secret),
+        client_secret: OAuth.openSecret(config.secret),
         redirect_uri: Meteor.absoluteUrl("_oauth/weibo?close", {replaceLocalhost: true}),
         grant_type: 'authorization_code'
       }});